[ecc] New module for Weierstrass curve config.
- Moved the Weierstrass elliptic curve config. class into its own
module, while retaining the `scriptless_zkp.ecc.ecc_utils` module for
ECC-related helper functions.

- Added helper functions to the `ecc_utils` module for generating random
nonce scalars and random nonce ECC point/scalar pairs, using parameters
from a provided Weierstrass elliptic curve config.

- Renamed the ECC-related exceptions module from `ecc_exceptions` to
simply `scriptless_zkp.ecc.exceptions`, since its package already
provides disambiguation with the `scriptless_zkp.exceptions` module.

GI-11
