+++ src/python/scriptless_zkp/ecc/ecc_utils.py
@@ -1,100 +1,44 @@
 """
-Support classes and helper functions for working with the elliptic curves used in elliptic curve cryptography (ECC).
+Helper functions for working with the elliptic curves used in elliptic curve cryptography (ECC).
 """
 from __future__ import annotations
 
-import math
-
-from attr import define
+from typing import Optional
 
 from Cryptodome.PublicKey import ECC
+from Cryptodome.Random import random
+
+from scriptless_zkp.ecc.weierstrass_curves import WeierstrassEllipticCurveConfig
 
 
-@define(slots=True, frozen=True)
-class WeierstrassEllipticCurveConfig:
+def generate_random_nonce(curve_config: WeierstrassEllipticCurveConfig) -> int:
     """
-    Configuration context for working with elliptic curves used in elliptic curve cryptography (ECC).
-    Support for obtaining a curve's base point (or generator) is provided, as well as frequently needed properties for
-    the order of the sub-group generated by this base point, the order of the field the elliptic curve points'
-    coordinates are over, and other intrinsic properties of a given elliptic curve.
+    Generates a random big integer in the range `[1, q]` inclusive, where `q` is the configured elliptic curve
+    sub-group's order (i.e., `|<G>|` where `<G>` is the sub-group generated by the public base point `G`).
+    :param curve_config: elliptic curve configuration to be used when generating the random private nonce scalar.
+    :return: a random big integer in the range `[1, q]` inclusive, where `q` is the configured elliptic curve
+             sub-group's order.
     """
+    return random.randint(1, curve_config.order)
 
-    curve: str
-    """A primary name used to identify this elliptic curve, which may be a commonly-used abbreviated form."""
 
-    order: int
-    """The order of the sub-group generated by the common, public base point (generator) of this elliptic curve."""
-
-    modulus: int
-    """The order of the field this elliptic curve points' coordinates are over."""
-
-    coeff_a: int
-    """This elliptic curve's coefficient `a`, in the Weierstrass elliptic curve equation: `y² = x³ + ax + b`"""
-
-    coeff_b: int
-    """This elliptic curve's coefficient `b`, in the Weierstrass elliptic curve equation: `y² = x³ + ax + b`"""
-
-    size_bits: int
+def generate_random_nonce_pair(
+        curve_config: WeierstrassEllipticCurveConfig,
+        ecc_base_point: Optional[ECC.EccPoint] = None
+) -> (int, ECC.EccPoint):
     """
-    Size of this elliptic curve's sub-group in bits, the bit-length of the order of the sub-group generated by the
-    curve's common, public base point (generator).
+    Generates a random private nonce scalar (big integer) and an associated nonce elliptic curve (EC) point, constructed
+    using the provided EC base point (as `nonce_point := nonce * base_point`), defaulting to the configured elliptic
+    curve's public generator `G` if an EC base point is not provided.
+    :param curve_config: elliptic curve configuration to be used when generating the random nonce EC point.
+    :param ecc_base_point: the ECC base point to be used for constructing the nonce EC point
+           (as `nonce_point := nonce * base_point`), defaulting to the configured elliptic curve sub-group's
+           generator `G` if not provided.
+    :return: (int, ECC.EccPoint) a randomly-generated nonce and an associated nonce EC point, constructed using
+             the provided EC base point (or the configured curve's public generator `G` if not provided).
     """
+    base_point: ECC.EccPoint = ecc_base_point if ecc_base_point else curve_config.base_point
+    private_nonce: int = generate_random_nonce(curve_config)  # random big integer in range [1, order] inclusive
+    nonce_point: ECC.EccPoint = base_point * private_nonce
 
-    curve_aliases: list[str]
-    """A list of one or more aliases this elliptic curve is known by, including the primary name given by `curve`."""
-
-    @property
-    def curve_size_bytes(self) -> int:
-        return math.ceil(self.size_bits / 8)
-
-    @property
-    def base_point(self) -> ECC.EccPoint:
-        """
-        A common, public base point (generator) of the elliptic curve, which generates a cyclic sub-group with order
-        ``self.order``.
-        """
-        return ECC.construct(curve=self.curve, d=1).pointQ
-
-    @property
-    def identity(self) -> ECC.EccPoint:
-        """
-        Returns the point-at-infinity (`O`), which is the elliptic curve group's additive identity.
-        """
-        return self.base_point.point_at_infinity()
-
-    @staticmethod
-    def is_identity(point: ECC.EccPoint) -> bool:
-        """
-        Returns whether the given elliptic curve point is the point-at-infinity (`O`), the group's additive identity.
-        """
-        return point.is_point_at_infinity()
-
-    def has_curve_name(self, curve_name: str):
-        """
-        Returns whether the given name matches a known alias of this Weierstrass elliptic curve configuration context.
-        """
-        return curve_name in self.curve_aliases
-
-    @classmethod
-    def secp256r1(cls) -> WeierstrassEllipticCurveConfig:
-        """
-        Factory function for constructing a configuration context for the standard NIST P-256 elliptic curve (also
-        known as SECP-256r1).
-        """
-        return WeierstrassEllipticCurveConfig(
-            curve="P-256",
-            order=0xffffffff00000000ffffffffffffffffbce6faada7179e84f3b9cac2fc632551,
-            modulus=0xffffffff00000001000000000000000000000000ffffffffffffffffffffffff,
-            coeff_a=0xffffffff00000001000000000000000000000000fffffffffffffffffffffffc,
-            coeff_b=0x5ac635d8aa3a93e7b3ebbd55769886bc651d06b0cc53b0f63bce3c3e27d2604b,
-            size_bits=256,
-            curve_aliases=["p256", "NIST P-256", "P-256", "prime256v1", "secp256r1", "nistp256"]
-        )
-
-    @classmethod
-    def p256(cls) -> WeierstrassEllipticCurveConfig:
-        """
-        Convenience factory function for constructing a configuration context for the standard NIST P-256 elliptic curve
-        (also known as SECP-256r1), named after a common short alias.
-        """
-        return cls.secp256r1()
+    return private_nonce, nonce_point
